requirements.txt
alabaster==0.7.10
astroid==1.6.2
Babel==2.5.3
bleach==2.1.3
certifi==2018.1.18
chardet==3.0.4
click==6.7
cloudpickle==0.5.2
decorator==4.2.1
docutils==0.14
entrypoints==0.2.3
Flask==0.12.2
html5lib==1.0.1
idna==2.6
imagesize==1.0.0
imutils==0.4.6
ipykernel==4.8.2
ipython==6.2.1
ipython-genutils==0.2.0
isort==4.3.4
itsdangerous==0.24
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
numpy==1.14.2
numpydoc==0.7.0
opencv-python==3.4.0.12
packaging==17.1
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
psutil==5.4.3
ptyprocess==0.5.2
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.8.3
PyOpenGL==3.1.0
pyparsing==2.2.0
PyQt5==5.9.2
python-dateutil==2.7.0
pytz==2018.3
pyzmq==17.0.0
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.4.0
requests==2.18.4
rope==0.10.7
simplegeneric==0.8.1
sip==4.19.8
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.2
sphinxcontrib-websupport==1.0.1
spyder==3.2.8
testpath==0.3.1
tornado==5.0.1
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
wrapt==1.10.11
